[CS] Rename ArrayOpenerNewlineFixer to ArrayOpenerAndCloserNewlineFixer
