+++ packages/coding-standard/tests/Fixer/ArrayNotation/ArrayOpenerAndCloserNewlineFixerTest/Fixture/also_closer.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerNewlineFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixerTest\Fixture;
 
 final class AlsoCloser
 {
@@ -17,7 +17,7 @@ final class AlsoCloser
 -----
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerNewlineFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixerTest\Fixture;
 
 final class AlsoCloser
 {
