+++ packages/coding-standard/tests/Fixer/ArrayNotation/ArrayOpenerAndCloserNewlineFixerTest/Fixture/simple_opener.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerNewlineFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixerTest\Fixture;
 
 final class SimpleOpener
 {
@@ -20,7 +20,7 @@ final class SimpleOpener
 -----
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerNewlineFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixerTest\Fixture;
 
 final class SimpleOpener
 {
