+++ packages/coding-standard/tests/Fixer/ArrayNotation/ArrayOpenerAndCloserNewlineFixerTest/Fixture/skip_double_opener.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerNewlineFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixerTest\Fixture;
 
 use PhpCsFixer\Fixer\ArrayNotation\ArraySyntaxFixer;
