+++ packages/easy-coding-standard/config/set/common/array.php
@@ -10,6 +10,7 @@ use PhpCsFixer\Fixer\ArrayNotation\TrimArraySpacesFixer;
 use PhpCsFixer\Fixer\ArrayNotation\WhitespaceAfterCommaInArrayFixer;
 use PhpCsFixer\Fixer\Whitespace\ArrayIndentationFixer;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symplify\CodingStandard\Fixer\ArrayNotation\ArrayOpenerAndCloserNewlineFixer;
 use Symplify\CodingStandard\Fixer\ArrayNotation\ArrayOpenerNewlineFixer;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
@@ -17,7 +18,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
 
     // array spacing
     $services->set(NoWhitespaceBeforeCommaInArrayFixer::class);
-    $services->set(ArrayOpenerNewlineFixer::class);
+    $services->set(ArrayOpenerAndCloserNewlineFixer::class);
     $services->set(ArrayIndentationFixer::class);
     $services->set(TrimArraySpacesFixer::class);
     $services->set(WhitespaceAfterCommaInArrayFixer::class);
